+++ harness/haskell/microformats2-parser/src/PreflightCheck.hs
@@ -1,0 +1,6 @@
+module Main where
+
+import Data.Microformats2.Parser
+
+main = do
+  return()